chore: Add next-translate
