+++ examples/with-next-translate/.gitignore
@@ -1,0 +1,1 @@
+pages/
