+++ examples/with-next-translate/i18n.json
@@ -1,0 +1,11 @@
+{
+  "allLanguages": ["en", "ca"],
+  "defaultLanguage": "en",
+  "currentPagesDir": "_pages",
+  "finalPagesDir": "pages",
+  "localesPath": "locales",
+  "pages": {
+    "*": ["common"],
+    "/": ["home"]
+  }
+}
