+++ examples/with-next-translate/locales/ca/common.json
@@ -1,0 +1,5 @@
+{
+  "title": "Demostració with-next-translate",
+  "powered": "Powered by",
+  "alt": "Vercel logo"
+}
