+++ examples/with-next-translate/locales/ca/home.json
@@ -1,0 +1,10 @@
+{
+  "catalan": "Català",
+  "change-catalan": "Canvia a la versió en català",
+  "change-english": "Canvia a la versió en anglès",
+  "description": "Comença editant",
+  "english": "Anglès",
+  "next-docs": "Troba informació detallada sobre l'API de Next.js.",
+  "plugin-docs": "Llegeix més sobre com fer servir next-translation.",
+  "title": "<0>Benvingut a <1></1></0>"
+}
