+++ examples/with-next-translate/locales/en/common.json
@@ -1,0 +1,5 @@
+{
+  "title": "Example with-next-translate",
+  "powered": "Powered by",
+  "alt": "Vercel logo"
+}
