+++ examples/with-next-translate/locales/en/home.json
@@ -1,0 +1,10 @@
+{
+  "catalan": "Catalan",
+  "change-catalan": "Change language to Catalan",
+  "change-english": "Change language to English",
+  "description": "Get started by editing",
+  "english": "English",
+  "next-docs": "Find in-depth information about Next.js features and API.",
+  "plugin-docs": "Read more about next-translation usage.",
+  "title": "<0>Welcome to <1></1></0>"
+}
